Added ACE support for different medical items
Added flexible ACE support for ranked medic system - It is now not only possible to get points when healing units using ACE, but also possible to distribute points for different medical actions and if the healer is medic or not. (Medic has to be set with Ace_medical_medicClass > 0 as unit variable. This also has a checkbox in 3DEN editor.