Fix race condition causing stat losswhen using missionProfileNamespace
This is an attempt to fix the occasional stat loss for players that are logged into the server during server launchup, e.g. when they forget to disconnect during a server restart. The issue seems to be the mission attempting to read missionProfileNamespace before it is loaded, which could return an empty d_player_hashmap and a subsqeuent call to saveMissionProfileNamespace that overwrites playerstats with zeros. By waiting for the mission profile to load and using a default empty hash map instead of a nil we avoid both the race condition and the accidental overwrite.

- the script now checks d_db_type and waits for isMissionProfileNamespaceLoaded before accessing the profile data
- missionProfileNamespace getVariable updated to provide a default empty HashMap instead of returning nil when no data was found

Please note that I did not get around to test it yet, if you want to test it yourself don't forget to backup your data before!